Allow getDitcher act to be called
This is part of the change to expose the MongoDB client in fh-db
to the user.

This allows a new act to be called without a type, called getDitcher
which will return the `ditcher.Ditcher` object created in the
`localdb` module. This will then expose the `Database` object and
the `mongo.Db` client within it.

Further changes are required in `fh-mbaas-api`.