🐛 RHMAP-15177 Fix for running locally when using v7 of fh-mbaas-api
The fh.sync API expects to be connecting directly to mongodb.
To support this, a change in fh-mbaas-api sets FH_MONGODB_CONN_URL
appropriately if FH_USE_LOCAL_DB is set. Therefore, there should be
no special behaviour here in fh-db, and it should just use the
value of FH_MONGODB_CONN_URL