🐛 RHMAP-16889 Allow mongodb connection string to work without auth
This addresses the following error when using fh-db (via fh-mbaas-api)
for local development i.e. when FH_USE_LOCAL_DB is true:

```
    LOCALDB error Error: Incorrect format for database connection string.
```

The actual error & stack is captured here.
https://gist.github.com/david-martin/eb96556f07ef05d8441238e70983d197
The mongo string parsing function is making the assumption that there is
an `@` symbol i.e. a user & password is specified before the mongodb
host.

The fix checks for the existense of the `@` symbol before tyring to
split the string and use parts of it