Fixing bug related to multiple restrictions on the same field
When adding multiple restrictions for the same field, e.g. "gt" and "lt" on a field containing a unix timestamp to search within a range, only the last restriction is used in the query. The restrictions for the field do not build up in the query object, instead they overwrite eachother, leaving only the last one active.